+++ examples/hosts/app-integration/schema-upgrade/src/modelVersion1/appModel.ts
