+++ examples/hosts/app-integration/schema-upgrade/src/modelVersion2/appModel.ts
